Added routers, added tags for documentation
