+++ app/routers/user.py
@@ -1,0 +1,37 @@
+from fastapi import Depends, status, HTTPException, APIRouter
+from app.utils import hash_password
+from app import models, schemas
+from app.database import get_db
+from sqlalchemy.orm import Session
+
+router = APIRouter(
+    prefix="/users",
+    tags=["Users"]
+)
+
+# CREATE USER
+@router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.UserResponse)
+def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
+
+    # hash the password - user.password
+    hashed_password = hash_password(user.password)
+    # replace the plain text password in user
+    user.password = hashed_password
+
+    new_user = models.User(**user.model_dump())
+    db.add(new_user)
+    db.commit()
+    db.refresh(new_user)
+
+    return new_user
+
+@router.get("/{id}", response_model=schemas.UserResponse)
+def get_user(id: int, db: Session = Depends(get_db)):
+    user = db.query(models.User).filter(models.User.id == id).first()
+
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"user with id: {id} not found"
+        )
+    return user
